🍟 [FIX] rename translation into content and make tag translation a separate file
