+++ src/modules/intelligence/domain/repositories/__init__.py
@@ -1,5 +1,6 @@
 from .analysis_repository import AnalysisRepository
-from .analyses_translation_repository import AnalysesTranslationRepository, TagTranslationRepository
+from .analyses_translation_repository import AnalysesTranslationRepository
+from .tag_translation_repository import TagTranslationRepository
 
 
 __all__ = [
